fix(lint): split ui main.rs and overview.rs under the 600-line gate
main.rs (663) and overview.rs (661) crept past the lowered 600-line
pre-push gate. Extracts the Header component into src/header.rs (used
only by Shell) and the NEEDS ATTENTION triage classification/table into
src/overview_attention.rs (used only by OverviewPage), following the
existing overview_recent_runs/overview_upcoming split convention. No
behavior changes, pure code move.
